# Conflicts:
#	ManagingDiet/target/m2e-wtp/web-resources/META-INF/maven/com.diet/biz/pom.properties

+++ ManagingDiet/src/main/java/com/diet/biz/dietProgram/impl/DietProgramServiceImpl.java
@@ -19,7 +19,7 @@ public class DietProgramServiceImpl implements DietProgramService{
 	private DietProgramDAOSpring dietProgramDAO;
 
 	@Override
-	public double dietStep1(KcalVO vo) {
+	public int dietStep1(KcalVO vo) {
 		return dietProgramDAO.dietStep1(vo);
 	}
